Broke the shader into more parts, sphereRad and horosphereSize are know uniforms, implemented VREffect as defined from hypVR
Shader now is composed of a globals file, math file, scene file and the
main file. Essentially in the future we can switch the math and scene
file to different setups so we can experiment in different spaces.
Currently sphereRad and horosphereSize are being set in
SceneManipulator, I wasn't certain of what values they needed to be so I
left the values the same for each different scenario. VREffect is
implemented as it was in the hypVR repository however I haven't been
able to test its actual vr capabilities yet.