Attempted shiny effect when light bounces from source off surface into camera
Trickiness from fact that you shouldn’t get shine from reflections off
the back side of the surface. This is solved with an if statement, but
we get an odd looking hard edge.  None of this would happen if we
actually marched along a ray to try to get to the source.